[Refactor](Other): Separated into modules and added a keymapping to activate.
