+++ lua/ds_omega/config/Navigation/other/keymappings.lua
@@ -1,0 +1,5 @@
+return {
+  n = {
+    ['<Leader>n6'] = { '<Cmd>Other<Cr>', 'Other file' },
+  }
+}
